Move BuildArtifact.stream() tests to test/regression/issue/10004.test.ts
BuildArtifact.stream() worked from its introduction (982dc0b441, May
2023) until #7650 (c8305df5c2, Dec 2023) added the cached-stream fast
path to Blob.getStream, so #10004 is a true regression and the test
belongs under test/regression/issue/. This also keeps the regression
test out of bun-build-api.test.ts, which is otherwise unchanged by
this PR.
